flask_app/requirements.txt
Flask==3.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.31.0
